+++ core/src/main/java/br/ufmg/dcc/latin/aspect/WeightedPassageAspectMining.java
@@ -18,12 +18,12 @@ public class WeightedPassageAspectMining extends AspectMining {
 	
 	public WeightedPassageAspectMining() {
 		n = RetrievalCache.docids.length;
-		importance = new float[0];
-		novelty = new float[0];
-		coverage = new float[n][0];
-		s = new float[0];
-		v = new float[0];
-		accumulatedRelevance = new float[0];
+		importance = new double[0];
+		novelty = new double[0];
+		coverage = new double[n][0];
+		s = new double[0];
+		v = new double[0];
+		accumulatedRelevance = new double[0];
 		flatAspectModel = new FlatAspectModel();
 	}
 
@@ -49,12 +49,12 @@ public class WeightedPassageAspectMining extends AspectMining {
 		if (aspectSize == 0) {
 			return;
 		}
-		importance = new float[aspectSize];
-		novelty = new float[aspectSize];
-		coverage = new float[n][aspectSize];
-		v = new float[aspectSize];
-		s = new float[aspectSize];
-		features = new float[n][aspectSize][];
+		importance = new double[aspectSize];
+		novelty = new double[aspectSize];
+		coverage = new double[n][aspectSize];
+		v = new double[aspectSize];
+		s = new double[aspectSize];
+		features = new double[n][aspectSize][];
 
 		
 		float uniformImportance = 1.0f/aspectSize;
@@ -68,22 +68,22 @@ public class WeightedPassageAspectMining extends AspectMining {
 		for (String aspectId : flatAspectModel.getAspects()) {
 			int s = flatAspectModel.getAspectComponents(aspectId).size();
 			for(int j = 0;j< n ;++j) {
-				features[j][i] = new float[s];
+				features[j][i] = new double[s];
 			}
 			int k = 0;
 			for (String aspectComponent: flatAspectModel.getAspectComponents(aspectId)) {
 				
-				float[] scores = null;
+				double[] scores = null;
 				if (RetrievalCache.passageCache.containsKey(aspectComponent)) {
 					scores = RetrievalCache.passageCache.get(aspectComponent);
 				} else {
 					scores = RetrievalController.rerankResults(RetrievalCache.docids, index , QueryParser.escape(aspectComponent));
 					RetrievalCache.passageCache.put(aspectComponent, scores);
 				}
-				float passageWeight = flatAspectModel.getAspectWeight(aspectId, aspectComponent);
+				double passageWeight = flatAspectModel.getAspectWeight(aspectId, aspectComponent);
 			    scores = scaling(scores);
 			    for(int j = 0;j< n ;++j) {
-			    	float score = scores[j]*passageWeight;
+			    	double score = scores[j]*passageWeight;
 			    	features[j][i][k] = score;
 			    	if (coverage[j][i] < score) {
 			    		coverage[j][i] = score;
@@ -126,11 +126,11 @@ public class WeightedPassageAspectMining extends AspectMining {
 		if (aspectSize == 0) {
 			return;
 		}
-		importance = new float[aspectSize];
-		novelty = new float[aspectSize];
-		coverage = new float[n][aspectSize];
-		accumulatedRelevance = new float[aspectSize];
-		features = new float[n][aspectSize][];
+		importance = new double[aspectSize];
+		novelty = new double[aspectSize];
+		coverage = new double[n][aspectSize];
+		accumulatedRelevance = new double[aspectSize];
+		features = new double[n][aspectSize][];
 		
 		float uniformImportance = 1.0f/aspectSize;
 		
@@ -143,12 +143,12 @@ public class WeightedPassageAspectMining extends AspectMining {
 			
 			int s = flatAspectModel.getAspectComponents(aspectId).size();
 			for(int j = 0;j< n ;++j) {
-				features[j][i] = new float[s];
+				features[j][i] = new double[s];
 			}
 			int k = 0;
 			for (String aspectComponent: flatAspectModel.getAspectComponents(aspectId)) {
 				
-				float[] scores = null;
+				double[] scores = null;
 				if (RetrievalCache.passageCache.containsKey(aspectComponent)) {
 					scores = RetrievalCache.passageCache.get(aspectComponent);
 				} else {
@@ -157,9 +157,9 @@ public class WeightedPassageAspectMining extends AspectMining {
 				}
 				
 			    scores = scaling(scores);
-			    float passageWeight = flatAspectModel.getAspectWeight(aspectId, aspectComponent);
+			    double passageWeight = flatAspectModel.getAspectWeight(aspectId, aspectComponent);
 			    for(int j = 0;j< n ;++j) {
-			    	float score = scores[j]*passageWeight;
+			    	double score = scores[j]*passageWeight;
 			    	features[j][i][k] = score;
 			    	if (coverage[j][i] < score) {
 			    		coverage[j][i] = score;
@@ -170,7 +170,7 @@ public class WeightedPassageAspectMining extends AspectMining {
 
 			for(int j = 0;j< n ;++j) {
 				if (this.feedbacks[j] != null) {
-					float score = this.feedbacks[j].getRelevanceAspect(aspectId);
+					double score = this.feedbacks[j].getRelevanceAspect(aspectId);
 					coverage[j][i] = score;
 					
 					if (score > 0) {
